requirements.txt
pyTelegramBotAPI>=4.2.0
pySerial>=3.5